refactor(config): add dotenv
- update gitignore
