+++ apps/api/internal/config/config.go
@@ -1,29 +1,37 @@
 package config
 
-import "os"
+import (
+	"errors"
+	"log"
+	"os"
 
-// Config holds configuration values.
-type Config struct {
-	Port       string
-	SQLitePath string
-	MongoURI   string
-	JWTSecret  string
-}
+	"github.com/joho/godotenv"
+)
 
-// LoadConfig reads config from environment variables with defaults.
 func LoadConfig() (*Config, error) {
+	if os.Getenv("GO_ENV") != "production" {
+		if err := godotenv.Load(); err != nil {
+			log.Println("No .env file found, proceeding with system environment variables")
+		}
+	}
+
 	cfg := &Config{
-		Port:       getEnv("PORT", "8080"),
-		SQLitePath: getEnv("SQLITE_PATH", "/data/sqlite/users.db"),
-		MongoURI:   getEnv("MONGO_URI", "mongodb://root:strongpassword@mongodb:27017"),
-		JWTSecret:  getEnv("JWT_SECRET", "supersecretkey"),
+		Port:       os.Getenv("PORT"),
+		SQLitePath: os.Getenv("SQLITE_PATH"),
+		MongoURI:   os.Getenv("MONGO_URI"),
+		JWTSecret:  os.Getenv("JWT_SECRET"),
+	}
+	if cfg.Port == "" {
+		return nil, errors.New("environment variable PORT is not set")
+	}
+	if cfg.SQLitePath == "" {
+		return nil, errors.New("environment variable SQLITE_PATH is not set")
+	}
+	if cfg.MongoURI == "" {
+		return nil, errors.New("environment variable MONGO_URI is not set")
+	}
+	if cfg.JWTSecret == "" {
+		return nil, errors.New("environment variable JWT_SECRET is not set")
 	}
 	return cfg, nil
-}
-
-func getEnv(key, fallback string) string {
-	if value, ok := os.LookupEnv(key); ok {
-		return value
-	}
-	return fallback
 }
