+++ apps/api/internal/config/model.go
@@ -1,0 +1,8 @@
+package config
+
+type Config struct {
+	Port       string
+	SQLitePath string
+	MongoURI   string
+	JWTSecret  string
+}
